+++ samples/bfs/bfs.vcxproj
@@ -184,7 +184,7 @@ copy "$(CudaToolkitBinDir)\cudart*.dll" "$(OutDir)"</Command>
     <ClInclude Include="..\..\include\groute\cta_work.h" />
     <ClInclude Include="..\..\include\groute\device_router.cuh" />
     <ClInclude Include="..\..\include\groute\event_pool.h" />
-    <ClInclude Include="..\..\include\groute\fused_distributed_worklist.h" />
+    <ClInclude Include="..\..\include\groute\distributed_worklist.h" />
     <ClInclude Include="..\..\include\groute\fused_worker.h" />
     <ClInclude Include="..\..\include\groute\graphs\common.h" />
     <ClInclude Include="..\..\include\groute\graphs\coo_graph.h" />
