+++ samples/bfs/bfs.vcxproj.filters
@@ -100,7 +100,7 @@
     <ClInclude Include="..\..\include\groute\policy.h">
       <Filter>groute</Filter>
     </ClInclude>
-    <ClInclude Include="..\..\include\groute\fused_distributed_worklist.h">
+    <ClInclude Include="..\..\include\groute\distributed_worklist.h">
       <Filter>groute</Filter>
     </ClInclude>
   </ItemGroup>
