+++ samples/sssp/sssp_async_opt.cu
@@ -35,7 +35,7 @@
 #include <gflags/gflags.h>
 
 #include <groute/event_pool.h>
-#include <groute/fused_distributed_worklist.h>
+#include <groute/distributed_worklist.h>
 #include <groute/cta_work.h>
 
 #include <groute/graphs/csr_graph.h>
